requirements.txt
APScheduler==3.5.1
cairocffi==0.8.0
CairoSVG==2.1.3
cffi==1.11.4
cssselect2==0.2.1
defusedxml==0.5.0
Django==2.0
django-annoying==0.10.4
django-apscheduler==0.2.5
django-background-tasks==1.1.13
django-compat==1.0.15
django-cors-headers==2.1.0
django-enumchoicefield==0.9.0
djangorestframework==3.7.4
gunicorn==19.7.1
html5lib==1.0b10
httplib2==0.10.3
mysqlclient==1.3.12
pdfrw==0.4
Pillow==5.0.0
pycparser==2.18
PyPDF2==1.26.0
Pyphen==0.9.4
python-dateutil==2.6.1
pytz==2017.3
reportlab==3.4.0
six==1.11.0
tinycss2==0.6.1
tzlocal==1.5.1
WeasyPrint==0.42.2
webencodings==0.5.1
xhtml2pdf==0.2.1